requirements.txt
Pyrogram~=2.0.59
aiohttp~=3.8.1
shortzy~=0.0.4
python-dotenv~=0.20.0
schedule